+++ capytaine/green_functions/delhommeau.py
@@ -6,14 +6,13 @@
 
 import logging
 from functools import lru_cache
+from importlib import import_module
 
 import numpy as np
 
 from capytaine.tools.prony_decomposition import exponential_decomposition, error_exponential_decomposition
 
 from capytaine.green_functions.abstract_green_function import AbstractGreenFunction
-import capytaine.green_functions.Delhommeau_f90 as Delhommeau_f90
-import capytaine.green_functions.XieDelhommeau_f90 as XieDelhommeau_f90
 
 LOG = logging.getLogger(__name__)
 
@@ -38,24 +37,30 @@ class Delhommeau(AbstractGreenFunction):
         The implementation of the Prony decomposition used to compute the finite depth Green function.
         Accepted values: :code:`'fortran'` for Nemoh's implementation (by default), :code:`'python'` for an experimental Python implementation.
         See :func:`find_best_exponential_decomposition`.
+    floating_point_precision: string, optional
+        Either :code:`'float32'` for single precision computations or :code:`'float64'` for double precision computations.
+        Default: :code:`'float64'`.
 
     Attributes
     ----------
-    tabulated_r_range: numpy.array of shape(tabulation_nr,)
-    tabulated_z_range: numpy.array of shape(tabulation_nz,)
+    tabulated_r_range: numpy.array of shape (tabulation_nr,) and type floating_point_precision
+    tabulated_z_range: numpy.array of shape (tabulation_nz,) and type floating_point_precision
         Coordinates of the tabulation points.
-    tabulated_integrals: numpy.array of shape (tabulation_nr, tabulation_nz, 2, 2)
+    tabulated_integrals: numpy.array of shape (tabulation_nr, tabulation_nz, 2, 2) and type floating_point_precision
         Tabulated Delhommeau integrals.
     """
 
-    fortran_core = Delhommeau_f90
+    fortran_core_basename = "Delhommeau"
 
     def __init__(self, *,
                  tabulation_nr=328,
                  tabulation_nz=46,
                  tabulation_nb_integration_points=251,
                  finite_depth_prony_decomposition_method='fortran',
+                 floating_point_precision='float64',
                  ):
+
+        self.fortran_core = import_module(f"capytaine.green_functions.libs.{self.fortran_core_basename}_{floating_point_precision}")
 
         self.tabulated_r_range = self.fortran_core.delhommeau_integrals.default_r_spacing(tabulation_nr)
         self.tabulated_z_range = self.fortran_core.delhommeau_integrals.default_z_spacing(tabulation_nz)
@@ -71,6 +76,7 @@ class Delhommeau(AbstractGreenFunction):
             'tabulation_nz': tabulation_nz,
             'tabulation_nb_integration_points': tabulation_nb_integration_points,
             'finite_depth_prony_decomposition_method': finite_depth_prony_decomposition_method,
+            'floating_point_precision': floating_point_precision,
         }
 
         self._hash = hash(self.exportable_settings.values())
@@ -112,7 +118,7 @@ class Delhommeau(AbstractGreenFunction):
             # The function that will be approximated.
             @np.vectorize
             def f(x):
-                return Delhommeau_f90.initialize_green_wave.ff(x, dimensionless_omega, dimensionless_wavenumber)
+                return self.fortran_core.initialize_green_wave.ff(x, dimensionless_omega, dimensionless_wavenumber)
 
             # Try different increasing number of exponentials
             for n_exp in range(4, 31, 2):
@@ -132,7 +138,7 @@ class Delhommeau(AbstractGreenFunction):
                             dimensionless_omega, dimensionless_wavenumber)
 
         elif self.finite_depth_prony_decomposition_method.lower() == 'fortran':
-            lamda, a, nexp = Delhommeau_f90.old_prony_decomposition.lisc(dimensionless_omega, dimensionless_wavenumber)
+            lamda, a, nexp = self.fortran_core.old_prony_decomposition.lisc(dimensionless_omega, dimensionless_wavenumber)
             lamda = lamda[:nexp]
             a = a[:nexp]
 
@@ -226,4 +232,4 @@ class XieDelhommeau(Delhommeau):
     Same arguments and methods as :class:`Delhommeau`.
     """
 
-    fortran_core = XieDelhommeau_f90
+    fortran_core_basename = "XieDelhommeau"